Feature vectors turned to short for efficiency

+++ src/main/java/knn/classification/ClassificationAlgorithm.java
@@ -6,5 +6,5 @@ public interface ClassificationAlgorithm {
      *   Input - vector of features
      *   Output - vector of class predictions for each of output classes
      */
-    int[] classify(double[] vector);
+    int[] classify(short[] vector);
 }
