+++ src/main/java/knn/similarity/SimilarityMeasure.java
@@ -1,5 +1,5 @@
 package knn.similarity;
 
 public interface SimilarityMeasure {
-    double calculateSimilarity(double[] v1, double[] v2);
+    double calculateSimilarity(short[] v1, short[] v2);
 }
